pyproject.toml
[tool.poetry]
name = "mgeo"
version = "0.1.0"
description = ""
authors = ["mocobt <mocobt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pillow = "6.2.0"
numpy = "^1.18.1"
matplotlib = "^3.1.3"
sklearn = "^0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
